requirements.txt
blinker==1.9.0
click==8.2.0
flask==3.1.0
itsdangerous==2.2.0
jinja2==3.1.6
logging==0.4.9.6
markupsafe==3.0.2
pathlib==1.0.1
pillow==11.2.1
tomli==2.2.1
werkzeug==3.1.3
